+++ docs/src/release_notes.rst
@@ -6,6 +6,14 @@ Release Notes
 0.4.x
 =====
 
+0.4.10
+------
++ Added :py:meth:`~tcex.tcex_resources.Resource.victims` method to :py:mod:`~tcex.tcex_resources.Resource` module.
++ Added :py:meth:`~tcex.tcex_resources.Resource.victim_assets` method to :py:mod:`~tcex.tcex_resources.Resource` module.
++ Added :py:meth:`~tcex.tcex_resources.Indicator.observation` methods to :py:mod:`~tcex.tcex_resources.Resource` module.
++ Added :py:meth:`~tcex.tcex_resources.Indicator.observation_count` methods to :py:mod:`~tcex.tcex_resources.Resource` module.
++ Added :py:meth:`~tcex.tcex_resources.Indicator.observed` methods to :py:mod:`~tcex.tcex_resources.Resource` module.
+
 0.4.9
 -----
 + Added :py:meth:`~tcex.tcex_resources.Signature.download` method to download signature data.